streamlit_gui/requirements.txt
streamlit
plotly
yfinance
numpy
pandas
google-cloud-storage